+++ tests/api/api/test_functions.py
@@ -37,6 +37,7 @@ import server.api.main
 import server.api.utils.builder
 import server.api.utils.clients.chief
 import server.api.utils.clients.iguazio
+import server.api.utils.functions
 import server.api.utils.singletons.db
 import server.api.utils.singletons.k8s
 import tests.api.api.utils
@@ -85,6 +86,89 @@ def test_build_status_pod_not_found(
         },
     )
     assert response.status_code == HTTPStatus.NOT_FOUND.value
+
+
+@pytest.mark.asyncio
+async def test_list_functions_with_pagination(
+    db: sqlalchemy.orm.Session, async_client: httpx.AsyncClient
+):
+    """
+    Test list functions with pagination.
+    Create 25 functions, request the first page, then use token to request 2nd and 3rd pages.
+    3rd page will contain only 5 functions instead of 10.
+    The 4th request with the token will return 404 as the token is now expired.
+    Requesting the 4th page without token will return 0 functions.
+    """
+    await tests.api.api.utils.create_project_async(async_client, PROJECT)
+
+    number_of_functions = 25
+    page_size = 10
+    for counter in range(number_of_functions):
+        function_name = f"function-name-{counter}"
+        function = {
+            "kind": "job",
+            "metadata": {
+                "name": function_name,
+                "project": "project-name",
+                "tag": "function-tag",
+            },
+            "spec": {"image": "mlrun/mlrun"},
+        }
+
+        post_function_response = await async_client.post(
+            f"projects/{PROJECT}/functions/{function_name}",
+            json=function,
+        )
+
+        assert post_function_response.status_code == HTTPStatus.OK.value
+
+    response = await async_client.get(
+        f"projects/{PROJECT}/functions",
+        params={
+            "page": 1,
+            "page-size": page_size,
+        },
+    )
+    _assert_pagination_info(response, 1, page_size, page_size, "function-name-0")
+    page_token = response.json()["pagination"]["page-token"]
+
+    response = await async_client.get(
+        f"projects/{PROJECT}/functions",
+        params={
+            "page-token": page_token,
+        },
+    )
+    _assert_pagination_info(response, 2, page_size, page_size, "function-name-10")
+
+    response = await async_client.get(
+        f"projects/{PROJECT}/functions",
+        params={
+            "page-token": page_token,
+        },
+    )
+    _assert_pagination_info(response, 3, 5, page_size, "function-name-20")
+
+    response = await async_client.get(
+        f"projects/{PROJECT}/functions",
+        params={
+            "page-token": page_token,
+        },
+    )
+    assert response.status_code == HTTPStatus.NOT_FOUND.value
+
+
+def _assert_pagination_info(
+    response,
+    expected_page,
+    expected_results_count,
+    expected_page_size,
+    expected_first_result_name,
+):
+    assert response.status_code == HTTPStatus.OK.value
+    assert response.json()["pagination"]["page"] == expected_page
+    assert response.json()["pagination"]["page-size"] == expected_page_size
+    assert len(response.json()["funcs"]) == expected_results_count
+    assert response.json()["funcs"][0]["metadata"]["name"] == expected_first_result_name
 
 
 @pytest.mark.asyncio
@@ -379,7 +463,7 @@ def test_tracking_on_serving(
 
     functions_to_monkeypatch = {
         server.api.api.utils: ["apply_enrichment_and_validation_on_function"],
-        server.api.api.endpoints.functions: [
+        server.api.utils.functions: [
             "process_model_monitoring_secret",
             "create_model_monitoring_stream",
         ],
@@ -433,14 +517,14 @@ def test_build_function_with_mlrun_bool(
             "tag": "latest",
         },
     }
-    original_build_function = server.api.api.endpoints.functions._build_function
+    original_build_function = server.api.utils.functions.build_function
     for with_mlrun in [True, False]:
         request_body = {
             "function": function_dict,
             "with_mlrun": with_mlrun,
         }
         function = mlrun.new_function(runtime=function_dict)
-        server.api.api.endpoints.functions._build_function = unittest.mock.Mock(
+        server.api.utils.functions.build_function = unittest.mock.Mock(
             return_value=(function, True)
         )
         response = client.post(
@@ -448,11 +532,8 @@ def test_build_function_with_mlrun_bool(
             json=request_body,
         )
         assert response.status_code == HTTPStatus.OK.value
-        assert (
-            server.api.api.endpoints.functions._build_function.call_args[0][3]
-            == with_mlrun
-        )
-    server.api.api.endpoints.functions._build_function = original_build_function
+        assert server.api.utils.functions.build_function.call_args[0][3] == with_mlrun
+    server.api.utils.functions.build_function = original_build_function
 
 
 @pytest.mark.parametrize(
